Allow loading .ply with loadFileAsync
